made input and label set to a component and added document for jest
