+++ src/components/__tests__/CooniBtn.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { CooniBtn } from '../shared/CooniBtn';
+import { CooniBtn } from '@shared/CooniBtn';
 
 // const component = shallow(
 //     <CooniBtn white={true} btnTxt='CooniBtn 1st test' />
@@ -10,7 +10,7 @@ import { CooniBtn } from '../shared/CooniBtn';
 // test for the pure component
 describe('CooniBtn shared component test', () => {
   const component = shallow(
-    <CooniBtn white={true} btnTxt='CooniBtn 2nd test' />
+    <CooniBtn white={true} btnTxt='CooniBtn 2nd test' />,
   );
 
   it('component and snapshot matches', () => {
