+++ src/components/__tests__/SignUp.test.tsx
@@ -10,16 +10,9 @@ describe('SignUp page DOM rendering test', () => {
   let page;
   const testStore = new Store();
   beforeEach(() => {
-    // const testStore = new Store();
     page = shallow(
         <SignUp test={true} store={testStore} />,
     );
-  });
-
-  it('the component input event is triggered with a mock argument and the state will be changed', () => {
-    const element = page.find('[id="emailInput"]');
-    element.simulate('change', { target: {id: 'emailInput', value: 'email' }});
-    expect(page.state().idDone).toEqual(true);
   });
 
   it('testStore is sent as props and its user property has to be defined', () => {
